Question 1
In this example, the getData function uses the fetch method to make a GET request to the https://jsonplaceholder.typicode.com/posts endpoint, which returns a Promise that resolves to a Response object.

We then use the json method of the Response object to extract the JSON data from the response, which also returns a Promise that resolves to the parsed JSON data.

Finally, we log the JSON data to the console. If an error occurs during the fetch request, we catch the error and log it to the console.

You can customize this example to fetch data from other endpoints of the JSON Placeholder API by replacing the URL used in the fetch method.